Fix flanker.utils circular imports
