+++ flanker/mime/message/utils.py
@@ -1,6 +1,9 @@
 from cStringIO import StringIO
 from contextlib import closing
 from email.generator import Generator
+from flanker.mime.message import errors
+import cchardet
+import chardet
 
 
 def python_message_to_string(msg):
@@ -9,3 +12,63 @@ def python_message_to_string(msg):
         g = Generator(fp, mangle_from_=False)
         g.flatten(msg, unixfrom=False)
         return fp.getvalue()
+
+def _guess_and_convert_with(value, detector=cchardet):
+    """
+    Try to guess the encoding of the passed value with the provided detector
+    and decode it.
+
+    The detector is either chardet or cchardet module.
+    """
+    charset = detector.detect(value)
+
+    if not charset["encoding"]:
+        raise errors.DecodingError("Failed to guess encoding")
+
+    try:
+        value = value.decode(charset["encoding"], "replace")
+        return value
+
+    except (UnicodeError, LookupError) as e:
+        raise errors.DecodingError(str(e))
+
+def _guess_and_convert(value):
+    """
+    Try to guess the encoding of the passed value and decode it.
+
+    Uses cchardet to guess the encoding and if guessing or decoding fails, falls
+    back to chardet which is much slower.
+    """
+    try:
+        return _guess_and_convert_with(value, detector=cchardet)
+    except:
+        return _guess_and_convert_with(value, detector=chardet)
+
+
+def _make_unicode(value, charset=None):
+    if isinstance(value, unicode):
+        return value
+
+    charset = charset or "utf-8"
+    try:
+        value = value.decode(charset, "strict")
+    except (UnicodeError, LookupError):
+        value = _guess_and_convert(value)
+
+    return value
+
+
+def to_utf8(value, charset=None):
+    '''
+    Safely returns a UTF-8 version of a given string
+    >>> utils.to_utf8(u'hi')
+        'hi'
+    '''
+
+    value = _make_unicode(value, charset)
+
+    return value.encode("utf-8", "strict")
+
+
+def to_unicode(value, charset=None):
+    return _make_unicode(value, charset)
